requirements.txt
affine==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.12
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
attrs==25.1.0
backports.tarfile==1.2.0
bidict==0.23.1
bleach==6.2.0
bokeh==3.6.3
branca==0.8.1
cattrs==24.1.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
colorcet==3.1.0
Columnar==1.3.1
contourpy==1.3.1
cycler==0.12.1
docutils==0.21.2
executing==2.2.0
fastapi==0.115.8
folium==0.19.4
fonttools==4.55.8
frozenlist==1.5.0
geopandas==1.0.1
h11==0.14.0
holoviews==1.20.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
hvplot==0.11.2
icecream==2.1.4
idna==3.10
ifaddr==0.2.0
importlib_metadata==8.6.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.6.0
kiwisolver==1.4.8
linkify-it-py==2.0.3
loguru==0.7.3
Markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==3.0.2
matplotlib==3.10.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
more-itertools==10.6.0
multidict==6.1.0
narwhals==1.25.2
networkx==3.4.2
nicegui==2.11.0
numpy==2.2.2
orjson==3.10.15
packaging==24.2
pandas==2.2.3
panel==1.6.0
param==2.2.0
pillow==11.1.0
platformdirs==4.3.6
propcache==0.2.1
pscript==0.7.7
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyinaturalist==0.20.1
pyogrio==0.10.0
pyparsing==3.2.1
pyproj==3.7.0
pyrate-limiter==2.10.0
python-box==7.3.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.11.2
python-multipart==0.0.20
python-socketio==5.12.1
pytz==2025.1
pyviz_comms==3.0.4
pywin32-ctypes==0.2.3
PyYAML==6.0.2
rasterio==1.4.3
referencing==0.36.2
requests==2.32.3
requests-cache==1.2.1
requests-ratelimiter==0.7.0
rich==13.9.4
rpds-py==0.22.3
shapely==2.0.7
simple-spaced-repetition==0.2.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
timeago==1.0.14
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uc-micro-py==1.0.3
url-normalize==1.4.3
urllib3==2.3.0
uvicorn==0.34.0
vbuild==0.8.2
watchfiles==1.0.4
wcwidth==0.2.13
webencodings==0.5.1
websockets==14.2
win32_setctime==1.2.0
wsproto==1.2.0
xyzservices==2025.1.0
yarl==1.18.3
zipp==3.21.0
